Important improvements has been made
Simulation of Lennard-Johns substance has been created (its physical condition and . All functions somehow relative with it has been debugged.
Ultra-simple visualisation has been made via plotly library (Python)

+++ SPH.h
@@ -7,7 +7,7 @@ class Kernel
 {
 public:
 	size_t D;
-	size_t C = 0;
+	double C = 0;
 	
 	Kernel(size_t);
